% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagger.R
\name{bag_control}
\alias{bag_control}
\title{Controlling the bagging process}
\usage{
bag_control(var_imp = TRUE, oob = NULL, allow_parallel = TRUE,
  sampling = "none")
}
\arguments{
\item{var_imp}{A single logical: should variable importance scores be calculated?}

\item{oob}{A metric set created by \code{\link[yardstick:metric_set]{yardstick::metric_set()}} or \code{NULL}. If not
NULL, then the out-of-bag samples are used to estimate model performance.}

\item{allow_parallel}{A single logical: should the model fits be done in
parallel (even if a parallel \code{plan()} has been created)?}

\item{sampling}{Either "none" or "down". For classification only. The
training data, after bootstrapping, will be sampled down within each class
(with replacement) to the size of the smallest class.}
}
\value{
A list.
}
\description{
\code{bag_control()} can set options for ancillary aspects of the bagging process.
}
