% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_tree.R
\name{bag_tree}
\alias{bag_tree}
\alias{update.bag_tree}
\title{General Interface for Decision Tree Models}
\usage{
bag_tree(mode = "unknown", cost_complexity = 0, tree_depth = NULL, min_n = 2)

\method{update}{bag_tree}(
  object,
  parameters = NULL,
  cost_complexity = NULL,
  tree_depth = NULL,
  min_n = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", "regression", or
"classification".}

\item{cost_complexity}{A positive number for the the cost/complexity
parameter (a.k.a. \code{Cp}) used by CART models (\code{rpart} only).}

\item{tree_depth}{An integer for maximum depth of the tree.}

\item{min_n}{An integer for the minimum number of data points
in a node that are required for the node to be split further.}

\item{object}{A bagged tree model specification.}
}
\description{
\code{bag_tree()} is a way to generate a \emph{specification} of a model
before fitting and allows the model to be created using
different packages in R. The main arguments for the
model are:
\itemize{
\item \code{cost_complexity}: The cost/complexity parameter (a.k.a. \code{Cp})
used by CART models (\code{rpart} only).
\item \code{tree_depth}: The \emph{maximum} depth of a tree (\code{rpart}).
\item \code{min_n}: The minimum number of data points in a node
that are required for the node to be split further.
}
These arguments are converted to their specific names at the
time that the model is fit. Other options and argument can be
set using \code{set_engine()}. If left to their defaults
here (\code{NULL}), the values are taken from the underlying model
functions. If parameters need to be modified, \code{update()} can be used
in lieu of recreating the object from scratch.
}
\details{
The model can be created using the \code{fit()} function using the
following \emph{engines}:
\itemize{
\item \pkg{R}:  \code{"rpart"} (the default) or \code{"C5.0"} (classification only)
}

Note that, for \code{rpart} models, but \code{cost_complexity} and
\code{tree_depth} can be both be specified but the package will give
precedence to \code{cost_complexity}. Also, \code{tree_depth} values
greater than 30 \code{rpart} will give nonsense results on 32-bit
machines.
}
\examples{
bag_tree(tree_depth = 5) \%>\% set_mode("classification")
model <- bag_tree(cost_complexity = 10, min_n = 3)
model
update(model, cost_complexity = 1)
update(model, cost_complexity = 1, fresh = TRUE)
}
\seealso{
[\code{\link[=fit]{fit()}}
}
