% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{package}
\name{baguette-package}
\alias{baguette}
\alias{baguette-package}
\title{baguette: Efficient Model Functions for Bagging}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tree- and rule-based models can be bagged (\doi{10.1007/BF00058655}) using this package and their predictions equations are stored in an efficient format to reduce the model objects size and speed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://baguette.tidymodels.org}
  \item \url{https://github.com/tidymodels/baguette}
  \item Report bugs at \url{https://github.com/tidymodels/baguette/issues}
}

}
\author{
\strong{Maintainer}: Max Kuhn \email{max@posit.co} (\href{https://orcid.org/0000-0003-2402-136X}{ORCID})

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
