% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{nnet_imp_garson}
\alias{nnet_imp_garson}
\title{Garson Importance Scores for neural Networks}
\usage{
nnet_imp_garson(object)
}
\arguments{
\item{object}{A \code{model_fit} or \code{nnet} object}
}
\value{
A tibble.
}
\description{
Garson Importance Scores for neural Networks
}
\keyword{internal}
