% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baguette-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{var_imp}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{var_imp}}}
}}

