% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagger.R
\name{bagger}
\alias{bagger}
\alias{bagger.default}
\alias{bagger.data.frame}
\alias{bagger.matrix}
\alias{bagger.formula}
\alias{bagger.recipe}
\title{Bagging functions}
\usage{
bagger(x, ...)

\method{bagger}{default}(x, ...)

\method{bagger}{data.frame}(x, y, model = "CART", B = 10L,
  opt = NULL, var_imp = FALSE, oob = FALSE, extract = NULL, ...)

\method{bagger}{matrix}(x, y, model = "CART", B = 10L, opt = NULL,
  var_imp = FALSE, oob = FALSE, extract = NULL, ...)

\method{bagger}{formula}(formula, data, model = "CART", B = 10L,
  opt = NULL, var_imp = FALSE, oob = FALSE, extract = NULL, ...)

\method{bagger}{recipe}(x, data, model = "CART", B = 10L, opt = NULL,
  var_imp = FALSE, oob = FALSE, extract = NULL, ...)
}
\arguments{
\item{x}{A data frame, matrix, or recipe (depending on the method being used.)}

\item{...}{Optional arguments to pass to the \code{extract} function.}

\item{y}{A numeric or factor vector of outcomes. Categorical outcomes (i.e
classes) should be represented as factors, not integers.}

\item{model}{A single character value for the model being bagged. Possible
values are "CART", "MARS", "C5.0" (classification only), and "model_rules"
(regression only).}

\item{B}{A single integer greater than 1 for the maximum number of bootstrap
samples/ensemble members (some model fits might fail).}

\item{opt}{A named list (or NULL) of arguments to pass to the underlying
model function. A list of possible arguments per model are givien in Details.}

\item{var_imp}{A logical: should variable importance scores be calculated?}

\item{oob}{A logical: should out-of-bag performance metrics be calculated?
TODO Instead maybe NULL or metric set object?}

\item{extract}{A function (or NULL) that can extract model-related aspects
of each ensemble member. See Details below.}

\item{formula}{An object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted. Note that
this package does not support multivariate outcomes and that, if some
predictors are factors, dummy variables will not be created.}

\item{data}{A data frame containing the variables used in the formula or
recipe.}
}
\description{
General suite of bagging functions for several models.
}
\details{
TBD
}
\examples{
mars_reg <- bagger(Sepal.Width ~ ., data = iris, model = "MARS", var_imp = TRUE)

library(AmesHousing)
ames <- make_ames()
# These take a while:

\dontrun{
ames_mars <- bagger(log10(Sale_Price) ~ ., data = ames, model = "MARS", var_imp = TRUE)
ames_rules <- bagger(log10(Sale_Price) ~ ., data = ames, model = "model_rules", var_imp = TRUE)
}
}
