% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagger.R
\name{control_bag}
\alias{control_bag}
\title{Controlling the bagging process}
\usage{
control_bag(
  var_imp = TRUE,
  allow_parallel = TRUE,
  sampling = "none",
  reduce = TRUE
)
}
\arguments{
\item{var_imp}{A single logical: should variable importance scores be calculated?}

\item{allow_parallel}{A single logical: should the model fits be done in
parallel (even if a parallel \code{plan()} has been created)?}

\item{sampling}{Either "none" or "down". For classification only. The
training data, after bootstrapping, will be sampled down within each class
(with replacement) to the size of the smallest class.}

\item{reduce}{Should models be modified to reduce their size on disk?}
}
\value{
A list.
}
\description{
\code{control_bag()} can set options for ancillary aspects of the bagging process.
}
